Depreciated function, Testing for bug fix.
Depreciated 'mcp2515_start_transmit' function call as it's sole purpose was to call another function and return. 
Per http://forum.arduino.cc/index.php?topic=152145.msg1654634#msg1654634
After review of the code, this appears to be correct but I need to test this before I apply it to the master branch.