Modified CNF definition values
Audited all CNFx register definitions for baud-rates to confirm sample point location, SAM bit status, and SJW status.  Changed all baud rates so that SAM bit is set (triple sample), SJW is half or less than PS2, and the sample point is at 75% or better where possible.  This should help increase the reliability of the protocol controller.  This was done for 8, 16, and 20MHz crystals frequencies.